Fix new bug

+++ docs/build_docs.sh
@@ -4,14 +4,16 @@
 
 # Step 0: clean the existing apidoc rst files and any previous build folder.
 # Make sure to not delete any important .git info for the CI auto_publish action !
+# Note: "find" crashes if the directory does not exist ... which is the case for some of the
+# Github Actions, hence the if ...  then ... else below. fpavogt, 2020-12-14.
 rm -rf ./source/modules
-#rm -rf ./build/*.html
-#rm -rf ./build/_sources
-#rm -rf ./build/_images
-#rm -rf ./build/_static
-#rm -rf ./build/.buildinfo
-#rm -rf ./build/.doctrees
-find ./build -not \( -path '*.git*' -o -name '.nojekyll' -o -name '.-gitignore' \) -delete
+if [ -d "./build" ]
+then
+    echo "Smart cleaning of the ./build directory."
+    find ./build -not \( -path '*.git*' -o -name '.nojekyll' -o -name '.-gitignore' \) -delete
+else
+    echo "FYI: directory ./build does not exists."
+fi
 
 # Step 1: run autodoc to generate all the docstring rst files.
 # Force the rewrite of all of them to capture *all* the recent changes.
